+++ src/game/Direction.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+enum class Direction
+{
+    up,
+    down,
+    left,
+    right
+};
+
