+++ src/game/SimpleMap.hpp
@@ -1,122 +1,10 @@
 #include <tuple>
 #include <vector>
 
+#include "ExplosionRange.hpp"
+
 #include "glm/glm.hpp"
 #include "physics/PhysicsEngine.hpp"
-
-enum class Direction
-{
-    up,
-    down,
-    left,
-    right
-};
-
-template <typename T, Direction diretction>
-struct Distance
-{
-    Distance() = default;
-    constexpr explicit Distance(int d) : distance{d}
-    {
-    }
-
-    operator int() const
-    {
-        return distance;
-    }
-
-    Distance operator+=(Distance other)
-    {
-        distance += other.distance;
-        return *this;
-    }
-
-    Distance operator+(Distance other)
-    {
-        auto dist = *this;
-        return dist += other;
-    }
-
-private:
-    int distance;
-};
-
-using LeftDistance = Distance<int, Direction::left>;
-using RightDistance = Distance<int, Direction::right>;
-using UpDistance = Distance<int, Direction::up>;
-using DownDistance = Distance<int, Direction::down>;
-
-constexpr LeftDistance operator"" _left(unsigned long long int d)
-{
-    return LeftDistance{static_cast<int>(d)};
-}
-constexpr RightDistance operator"" _right(unsigned long long int d)
-{
-    return RightDistance{static_cast<int>(d)};
-}
-constexpr UpDistance operator"" _up(unsigned long long int d)
-{
-    return UpDistance{static_cast<int>(d)};
-}
-constexpr DownDistance operator"" _down(unsigned long long int d)
-{
-    return DownDistance{static_cast<int>(d)};
-}
-
-class ExplosionRange
-{
-public:
-    ExplosionRange(LeftDistance left, RightDistance right, UpDistance up,
-                   DownDistance down)
-        : leftDist{left}, rightDist{right}, upDist{up}, downDist{down}
-    {
-    }
-
-    LeftDistance left() const
-    {
-        return leftDist;
-    }
-
-    RightDistance right() const
-    {
-        return rightDist;
-    }
-
-    UpDistance up() const
-    {
-        return upDist;
-    }
-
-    DownDistance down() const
-    {
-        return downDist;
-    }
-
-    bool operator==(const ExplosionRange& other) const
-    {
-        return std::tie(leftDist, rightDist, upDist, downDist) ==
-               std::tie(other.leftDist, other.rightDist, other.upDist,
-                        other.downDist);
-    }
-
-private:
-    LeftDistance leftDist;
-    RightDistance rightDist;
-    UpDistance upDist;
-    DownDistance downDist;
-};
-
-std::ostream& operator<<(std::ostream& out, const ExplosionRange& range);
-std::ostream& operator<<(std::ostream& out, const ExplosionRange& range)
-{
-    out << "{";
-    out << "Left: " << range.left() << " ,";
-    out << "Right: " << range.right() << " ,";
-    out << "Up: " << range.up() << " ,";
-    out << "Down: " << range.down();
-    out << "}";
-    return out;
-}
 
 class WallPositionsGenerator
 {
